+++ app/package.json
@@ -9,6 +9,7 @@
         "@ionic/react": "^5.6.10",
         "@ionic/react-router": "^5.6.10",
         "@kiwi/api": "workspace:*",
+        "@kiwi/models": "workspace:*",
         "@kiwi/ui": "workspace:*",
         "chartist": "^0.11.4",
         "classnames": "^2.3.1",
