+++ packages/api/package.json
@@ -7,12 +7,5 @@
     "devDependencies": {
         "rimraf": "^3.0.2",
         "typescript": "4.3.4"
-    },
-    "peerDependencies": {
-        "@ionic/react": "*",
-        "classnames": "*",
-        "react": "*",
-        "react-dom": "*",
-        "react-jss": "*"
     }
 }
