+++ rider-app/package.json
@@ -9,6 +9,7 @@
         "@ionic/react": "^5.6.10",
         "@ionic/react-router": "^5.6.10",
         "@kiwi/api": "workspace:*",
+        "@kiwi/models": "workspace:*",
         "@kiwi/ui": "workspace:*",
         "classnames": "^2.3.1",
         "date-fns": "^2.22.1",
