+++ rider-app/src/pages/profile.tsx
@@ -29,7 +29,7 @@ import {
 import {Typography} from '@kiwi/ui';
 import {useAuth} from '../contexts/auth';
 import FormUser from '../components/form-user';
-import {User} from '../models';
+import {User} from '@kiwi/models/rider';
 import FormRider from '../components/form-rider';
 import FormPassword from '../components/form-password';
 import kiwiApi from '../api';
